requirements.txt
# Application
fastapi==0.132.0
uvicorn==0.30.1
python-dotenv==1.2.1
anyio==4.12.1
pydantic==2.12.5

# LLM orchestrator
llm-orchestrator @ git+https://github.com/JeronimoFSNeto/llm-orchestrator.git@888c33e6d42d8f9821a6fe22ab075e28385b4329

# Testing
httpx==0.28.1
pytest==9.0.2
pytest-asyncio==1.3.0
